move test support to utils directory and update imports
